refactor: clean up launch common and add tests
