+++ src/util/process_life_cycle.py
@@ -6,7 +6,7 @@ from _signal import SIGABRT, SIGILL, SIGSEGV, SIGTERM
 from enum import Enum, auto
 from time import sleep
 
-from Constants import VERSION, RunMode
+from Constants import VERSION, RunMode, LINER. BUF_SIZE
 from NetworkConfiguration import init_network_config
 from calc.service_fee_calculator import ServiceFeeCalculator
 from cli.client_manager import ClientManager
@@ -21,8 +21,6 @@ from log_config import main_logger, init, verbose_logger
 from plugins import plugins
 
 logger = main_logger
-LINER = "--------------------------------------------"
-BUF_SIZE = 50
 
 
 class TrdState(Enum):
